@@ -7,7 +7,7 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -19,12 +19,12 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
-  shared_preferences: ^0.5.10
-  hive_flutter: ^0.3.1
-  dio: ^3.0.10
-  uuid: ^2.2.2
-  http: ^0.12.2
+  cupertino_icons: ^1.0.3
+  shared_preferences: ^2.0.5
+  hive_flutter: ^1.0.0
+  dio: ^4.0.0
+  uuid: ^3.0.4
+  http: ^0.13.3
 
 dev_dependencies:
   flutter_test:
