@@ -1,23 +1,23 @@
-import 'interfaces/i_auth_repository.dart';
-import 'package:flutter/foundation.dart';
+// import 'interfaces/i_auth_repository.dart';
+// import 'package:flutter/foundation.dart';
 
-import '../domain/entities/user.dart';
+// import '../domain/entities/user.dart';
 
-class AuthService {
-  final IAuthRepository authRepository;
+// class AuthService {
+//   final IAuthRepository authRepository;
 
-  AuthService({@required this.authRepository});
+//   AuthService({@required this.authRepository});
 
-  Future<User> signUp(String email, String password) async {
-    return await authRepository.signUp(email, password);
-  }
+//   Future<User> signUp(String email, String password) async {
+//     return await authRepository.signUp(email, password);
+//   }
 
-  Future<User> login(String email, String password) async {
-    return await authRepository.login(email, password);
-  }
+//   Future<User> login(String email, String password) async {
+//     return await authRepository.login(email, password);
+//   }
 
-  User logout() {
-    authRepository.logout();
-    return UnsignedUser();
-  }
-}
+//   User logout() {
+//     authRepository.logout();
+//     return UnsignedUser();
+//   }
+// }
