@@ -22,9 +22,6 @@ void main() async {
   });
 
   testWidgets('Check form validation logic', (tester) async {
-    //find an active RaisedButton.
-    final passwordBtn =
-        find.byWidgetPredicate((e) => e is RaisedButton && e.enabled);
     await tester.pumpWidget(App());
     await tester.pumpAndSettle();
     //
@@ -33,14 +30,17 @@ void main() async {
     //Enter some invalid email
     await tester.enterText(find.byKey(Key('__EmailField__')), 'email');
     await tester.pump();
+    //Not validated yet as it will validate on lose focus
+    expect(find.text('Enter a valid email'), findsNothing);
+    //Lose focus of the email field
+    await tester.enterText(find.byKey(Key('__PasswordField__')), '');
+    await tester.pump();
     //We expect to see a validation error
     expect(find.text('Enter a valid email'), findsOneWidget);
     //Enter a valid email
     await tester.enterText(find.byKey(Key('__EmailField__')), 'email@mail.com');
     await tester.pump();
     expect(find.text('Enter a valid email'), findsNothing);
-    //Raised button is not active
-    expect(passwordBtn, findsNothing);
 
     //
     //Enter some invalid password
@@ -52,16 +52,18 @@ void main() async {
     await tester.enterText(find.byKey(Key('__PasswordField__')), '12345test');
     await tester.pump();
     expect(find.text('Enter a valid password'), findsNothing);
-    //Raised button is  active
-    expect(passwordBtn, findsOneWidget);
+
     //
     //Enter some invalid password
     await tester.enterText(find.byKey(Key('__PasswordField__')), '12345');
     await tester.pump();
     //We expect to see a validation error
     expect(find.text('Enter a valid password'), findsOneWidget);
-    //Raised button is not active
-    expect(passwordBtn, findsNothing);
+
+    //Enter a valid password
+    await tester.enterText(find.byKey(Key('__PasswordField__')), '12345test');
+    await tester.pump();
+    expect(find.text('Enter a valid password'), findsNothing);
   });
 
   testWidgets('Should Sign up and log out manually', (tester) async {
@@ -76,20 +78,32 @@ void main() async {
     await tester.tap(find.byType(Checkbox));
     await tester.pump();
     //Sign up mode is activated
-    expect(find.text('Register'), findsOneWidget);
+    expect(find.text('Sign up'), findsOneWidget);
 
     //Enter a valid email and password
     await tester.enterText(find.byKey(Key('__EmailField__')), 'user1@mail.com');
     await tester.enterText(find.byKey(Key('__PasswordField__')), '12345user1');
     await tester.pump();
-    //Raise button is active
-    await tester.tap(find.byType(RaisedButton));
+    //Enter some invalid confirmation password
+    await tester.enterText(
+        find.byKey(Key('__ConfirmPasswordField__')), '12345');
+    await tester.pump();
+    //We expect to see a validation error
+    expect(find.text('Passwords do not match'), findsOneWidget);
+
+    //Enter a valid confirmation password
+    await tester.enterText(
+        find.byKey(Key('__ConfirmPasswordField__')), '12345user1');
+    await tester.pump();
+    expect(find.text('Passwords do not match'), findsNothing);
+
+    await tester.tap(find.byType(ElevatedButton));
     await tester.pump();
 
     //Expect to see a CircularProgressIndicator
     expect(find.byType(CircularProgressIndicator), findsOneWidget);
-    //RaisedButton is hidden
-    expect(find.byType(RaisedButton), findsNothing);
+    //ElevatedButton is hidden
+    expect(find.byType(ElevatedButton), findsNothing);
     storage.isAsyncRead = true;
 
     //After one second
@@ -100,11 +114,11 @@ void main() async {
     //
     //Check that the signed user is persisted
     expect(
-      storage.store['__UserToken__'].contains('"email":"user1@mail.com"'),
+      storage.store['__UserToken__']!.contains('"email":"user1@mail.com"'),
       isTrue,
     );
     expect(
-        storage.store['__UserToken__']
+        storage.store['__UserToken__']!
             .contains('"token":"token_user1@mail.com'),
         isTrue);
     //
@@ -135,13 +149,13 @@ void main() async {
     await tester.enterText(find.byKey(Key('__PasswordField__')), '12345user1');
     await tester.pump();
     //Raise button is active
-    await tester.tap(find.byType(RaisedButton));
+    await tester.tap(find.byType(ElevatedButton));
     await tester.pump();
 
     //Expect to see a CircularProgressIndicator
     expect(find.byType(CircularProgressIndicator), findsOneWidget);
-    //RaisedButton is hidden
-    expect(find.byType(RaisedButton), findsNothing);
+    //ElevatedButton is hidden
+    expect(find.byType(ElevatedButton), findsNothing);
     storage.isAsyncRead = true;
     //After one second
     await tester.pumpAndSettle(Duration(seconds: 1));
@@ -150,11 +164,11 @@ void main() async {
     //
     //Check that the signed user is persisted
     expect(
-      storage.store['__UserToken__'].contains('"email":"user1@mail.com"'),
+      storage.store['__UserToken__']!.contains('"email":"user1@mail.com"'),
       isTrue,
     );
     expect(
-        storage.store['__UserToken__']
+        storage.store['__UserToken__']!
             .contains('"token":"token_user1@mail.com"'),
         isTrue);
     //
