@@ -1,25 +1,26 @@
 import 'dart:convert';
 
+import 'package:http/http.dart' as http;
 import 'package:states_rebuilder/states_rebuilder.dart';
 
 import '../domain/common/extensions.dart';
+import '../domain/entities/user.dart';
 import '../domain/value_object/token.dart';
 import '../service/exceptions/auth_exception.dart';
-import 'package:http/http.dart' as http;
-
-import 'my_project_data.dart' as myProjectData; //TODO Delete this.
-import '../domain/entities/user.dart';
-import '../service/interfaces/i_auth_repository.dart';
+import 'my_project_data.dart' as myProjectData;
 
 //Go to https://console.firebase.google.com/project/YOUR_PROJECT_NAME/settings/general and get `webApiKey`. This will be your `webApiKey` const.
 const webApiKey = myProjectData.webApiKey; //TODO Use yours
 
-class FireBaseAuth implements IAuth<User, UserParam> {
+class FireBaseAuth implements IAuth<User?, UserParam> {
   @override
   Future<void> init() async {}
 
   @override
-  Future<User> signIn(UserParam param) {
+  Future<User?> signIn(UserParam? param) async {
+    if (param == null) {
+      return null;
+    }
     return _authenticate(
       param.email,
       param.password,
@@ -28,7 +29,10 @@ class FireBaseAuth implements IAuth<User, UserParam> {
   }
 
   @override
-  Future<User> signUp(UserParam param) {
+  Future<User?> signUp(UserParam? param) async {
+    if (param == null) {
+      return null;
+    }
     return _authenticate(
       param.email,
       param.password,
@@ -37,15 +41,15 @@ class FireBaseAuth implements IAuth<User, UserParam> {
   }
 
   @override
-  Future<void> signOut(UserParam param) async {}
+  Future<void> signOut(UserParam? param) async {}
 
-  Future<User> _authenticate(
+  Future<User?> _authenticate(
       String email, String password, String urlSegment) async {
     final url =
         'https://www.googleapis.com/identitytoolkit/v3/relyingparty/$urlSegment?key=$webApiKey';
     try {
       var response = await http.post(
-        url,
+        Uri.parse(url),
         body: json.encode(
           {
             'email': email,
@@ -56,22 +60,42 @@ class FireBaseAuth implements IAuth<User, UserParam> {
       );
       final responseData = json.decode(response.body);
       if (responseData['error'] != null) {
+        if (responseData['error']['message'] == 'INVALID_EMAIL') {
+          throw EmailException.invalidEmail();
+        }
+
+        if (responseData['error']['message'] == 'EMAIL_NOT_FOUND') {
+          throw EmailException.emailNotFound();
+        }
+        if (responseData['error']['message'] == 'EMAIL_EXISTS') {
+          throw EmailException.emailExists();
+        }
+
+        if (responseData['error']['message'] == 'INVALID_PASSWORD') {
+          throw PasswordException.invalidPassword();
+        }
+
+        if (responseData['error']['message'] == 'WEAK_PASSWORD') {
+          throw PasswordException.weakPassword();
+        }
+
         throw AuthException(responseData['error']['message']);
       }
 
       return User(
-          userId: responseData['localId'],
-          email: email,
-          token: Token(
-            token: responseData['idToken'],
-            expiryDate: DateTimeX.current.add(
-              Duration(
-                seconds: int.parse(
-                  responseData['expiresIn'],
-                ),
+        userId: responseData['localId'],
+        email: email,
+        token: Token(
+          token: responseData['idToken'],
+          expiryDate: DateTimeX.current.add(
+            Duration(
+              seconds: int.parse(
+                responseData['expiresIn'],
               ),
             ),
-          ));
+          ),
+        ),
+      );
     } catch (error) {
       throw error;
     }
@@ -82,58 +106,3 @@ class FireBaseAuth implements IAuth<User, UserParam> {
     // TODO: implement dispose
   }
 }
-
-// class FireBaseAuth implements IAuthRepository {
-//   @override
-//   Future<User> login(String email, String password) {
-//     return _authenticate(email, password, 'verifyPassword');
-//   }
-
-//   @override
-//   Future<User> signUp(String email, String password) {
-//     return _authenticate(email, password, 'signupNewUser');
-//   }
-
-//   @override
-//   Future<void> logout() async {
-//     //
-//   }
-
-//   Future<User> _authenticate(
-//       String email, String password, String urlSegment) async {
-//     final url =
-//         'https://www.googleapis.com/identitytoolkit/v3/relyingparty/$urlSegment?key=$webApiKey';
-//     try {
-//       var response = await http.post(
-//         url,
-//         body: json.encode(
-//           {
-//             'email': email,
-//             'password': password,
-//             'returnSecureToken': true,
-//           },
-//         ),
-//       );
-//       final responseData = json.decode(response.body);
-//       if (responseData['error'] != null) {
-//         throw AuthException(responseData['error']['message']);
-//       }
-
-//       return User(
-//           userId: responseData['localId'],
-//           email: email,
-//           token: Token(
-//             token: responseData['idToken'],
-//             expiryDate: DateTimeX.current.add(
-//               Duration(
-//                 seconds: int.parse(
-//                   responseData['expiresIn'],
-//                 ),
-//               ),
-//             ),
-//           ));
-//     } catch (error) {
-//       throw error;
-//     }
-//   }
-// }
