@@ -1,8 +1,6 @@
 import 'package:flutter/foundation.dart';
 import 'package:uuid/uuid.dart';
 
-import '../exceptions/validation_exception.dart';
-
 @immutable
 class Todo {
   final String id;
@@ -10,13 +8,10 @@ class Todo {
   final String note;
   final String task;
 
-  Todo(this.task, {String id, this.note, this.complete = false})
+  Todo(this.task, {String? id, this.note = '', this.complete = false})
       : id = id ?? Uuid().v4();
 
-  factory Todo.fromJson(Map<String, Object> map) {
-    if (map == null) {
-      return null;
-    }
+  factory Todo.fromJson(Map<String, dynamic> map) {
     return Todo(
       map['task'] as String,
       id: map['id'] as String,
@@ -26,8 +21,7 @@ class Todo {
   }
 
   // toJson is called just before persistance.
-  Map<String, Object> toJson() {
-    _validation();
+  Map<String, dynamic> toJson() {
     return {
       'complete': complete,
       'task': task,
@@ -36,21 +30,11 @@ class Todo {
     };
   }
 
-  void _validation() {
-    if (id == null) {
-      // Custom defined error classes
-      throw ValidationException('This todo has no ID!');
-    }
-    if (task == null || task.isEmpty) {
-      throw ValidationException('Empty task are not allowed');
-    }
-  }
-
   Todo copyWith({
-    String task,
-    String note,
-    bool complete,
-    String id,
+    String? task,
+    String? note,
+    bool? complete,
+    String? id,
   }) {
     return Todo(
       task ?? this.task,
