@@ -1,7 +1,7 @@
-import '../../domain/entities/user.dart';
+// import '../../domain/entities/user.dart';
 
-abstract class IAuthRepository {
-  Future<User> signUp(String email, String password);
-  Future<User> login(String email, String password);
-  Future<void> logout();
-}
+// abstract class IAuthRepository {
+//   Future<User> signUp(String email, String password);
+//   Future<User> login(String email, String password);
+//   Future<void> logout();
+// }
