@@ -1,57 +1,48 @@
 class EN {
-  String appTitle = 'States_rebuilder Example';
-  String todos = 'Todos';
-
-  String stats = 'Stats';
-
-  String showAll = 'Show All';
-
-  String showActive = 'Show Active';
-
-  String showCompleted = 'Show Completed';
-
-  String newTodoHint = 'What needs to be done?';
-
-  String markAllComplete = 'Mark all complete';
-
-  String markAllIncomplete = 'Mark all incomplete';
-
-  String clearCompleted = 'Clear completed';
-
-  String addTodo = 'Add Todo';
-
-  String editTodo = 'Edit Todo';
-
-  String saveChanges = 'Save changes';
-
-  String filterTodos = 'Filter Todos';
-
-  String deleteTodo = 'Delete Todo';
-
-  String todoDetails = 'Todo Details';
-
-  String emptyTodoError = 'Please enter some text';
-
-  String notesHint = 'Additional Notes...';
-
-  String completedTodos = 'Completed Todos';
-
-  String activeTodos = 'Active Todos';
-
-  String todoDeleted(String task) => 'Deleted "$task"';
-
-  String undo = 'Undo';
-
-  String deleteTodoConfirmation = 'Delete this todo?';
-
-  String delete = 'Delete';
-
-  String cancel = 'Cancel';
-
-  String switchToDarkMode = 'switch to dark mode';
-  String switchToLightMode = 'switch to light mode';
-
-  String systemLanguage = 'System language';
-
-  String logout = 'Logout';
+  final String email = 'Email';
+  final String password = 'Password';
+  final String confirmPassword = 'Confirm Password';
+  final String doNotHaveAnAccount = 'I do not have an account';
+  final String signIn = 'Sign in';
+  final String signUp = 'Sign up';
+  //
+  final String appTitle = 'States_rebuilder Example';
+  final String todos = 'Todos';
+  final String stats = 'Stats';
+  final String showAll = 'Show All';
+  final String showActive = 'Show Active';
+  final String showCompleted = 'Show Completed';
+  final String newTodoHint = 'What needs to be done?';
+  final String markAllComplete = 'Mark all complete';
+  final String markAllIncomplete = 'Mark all incomplete';
+  final String clearCompleted = 'Clear completed';
+  final String addTodo = 'Add Todo';
+  final String editTodo = 'Edit Todo';
+  final String saveChanges = 'Save changes';
+  final String filterTodos = 'Filter Todos';
+  final String deleteTodo = 'Delete Todo';
+  final String todoDetails = 'Todo Details';
+  final String emptyTodoError = 'Please enter some text';
+  final String notesHint = 'Additional Notes...';
+  final String completedTodos = 'Completed Todos';
+  final String activeTodos = 'Active Todos';
+  String todoDeleted(final String task) => 'Deleted "$task"';
+  final String undo = 'Undo';
+  final String deleteTodoConfirmation = 'Delete this todo?';
+  final String delete = 'Delete';
+  final String cancel = 'Cancel';
+  final String switchToDarkMode = 'switch to dark mode';
+  final String switchToLightMode = 'switch to light mode';
+  final String systemLanguage = 'System language';
+  final String logout = 'Logout';
+  //
+  final String enterValidPassword = 'Enter a valid password';
+  final String enterValidEmail = 'Enter a valid email';
+  final String invalidEmail = 'Non valid Email';
+  final String emailNotFound = 'Email Not found';
+  final String emailExists = 'Email already exists';
+  final String weakPassword = 'The password must be 6 characters long or more';
+  final String invalidPassword = 'Password is invalid';
+  final String pageNotFound = 'Page not found';
+  final String networkFailure = 'Network failure';
 }
