@@ -3,7 +3,7 @@ import 'package:states_rebuilder/states_rebuilder.dart';
 import 'package:ex_009_1_3_ca_todo_mvc_with_state_persistence_user_auth/domain/entities/todo.dart';
 
 class FakeTodosRepository implements ICRUD<Todo, String> {
-  Map<String, List<Todo>> todos;
+  late Map<String, List<Todo>> todos;
   dynamic error;
   FakeTodosRepository() {
     todos = {
@@ -13,39 +13,39 @@ class FakeTodosRepository implements ICRUD<Todo, String> {
     };
   }
   @override
-  Future<List<Todo>> read(String userId) async {
+  Future<List<Todo>> read(String? userId) async {
     await Future.delayed(Duration(seconds: 1));
     if (error != null) {
       throw error;
     }
-    return todos[userId];
+    return todos[userId]!;
   }
 
   @override
-  Future<dynamic> update(List<Todo> item, String userId) async {
+  Future<dynamic> update(List<Todo> item, String? userId) async {
     await Future.delayed(Duration(seconds: 1));
-    final index = todos[userId].indexWhere((e) => e.id == item.first.id);
-    todos[userId][index] = item.first;
+    final index = todos[userId]!.indexWhere((e) => e.id == item.first.id);
+    todos[userId]![index] = item.first;
     return true;
   }
 
   @override
-  Future<Todo> create(Todo item, String userId) async {
+  Future<Todo> create(Todo item, String? userId) async {
     await Future.delayed(Duration(seconds: 1));
     if (error != null) {
       throw error;
     }
-    todos[userId].add(item);
+    todos[userId]!.add(item);
     return item;
   }
 
   @override
-  Future<dynamic> delete(List<Todo> item, String userId) async {
+  Future<dynamic> delete(List<Todo> item, String? userId) async {
     if (error != null) {
       await Future.delayed(Duration(seconds: 1));
       throw error;
     }
-    todos[userId].remove(item.first);
+    todos[userId]!.remove(item.first);
     return true;
   }
 
